cleanup evm data building
